requirements.txt
pydantic
pyyaml
ping3
paho-mqtt